requirements.txt
flask==0.12.2
Flask-MYSQL==1.4.0
flask_bootstrap==3.3.7.1
Flask_SQLAlchemy==2.3.2
Flask-Login==0.4.1
Flask-WTF==0.14.2
